% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{webPivotTable}
\alias{webPivotTable}
\title{Pivot table visualisation using https://webpivottable.com/}
\usage{
webPivotTable(data, width = NULL, height = NULL)
}
\arguments{
\item{data}{: data.frame}

\item{width}{: Width (optional, defaults to automatic sizing)}

\item{height}{: Height (optional, defaults to automatic sizing)}
}
\description{
Pivot table visualisation using https://webpivottable.com/
}
\examples{

\dontrun{
\donttest{
# work only in browser, not in RStudio viewer
w <- webPivotTable(iris)
w # view in real browser

htmlwidgets::saveWidget(w, file = "pivot.html")

# in shiny
 require(shiny)
 
 ui <- fluidPage(
   webPivotTableOutput("wpt1"),
   hr(),
   webPivotTableOutput("wpt2")
 )

 server <- function(input, output, session) {
   output$wpt1 <- renderWebPivotTable({
      webPivotTable(iris)
   })

   output$wpt2 <- renderWebPivotTable({
      webPivotTable(mtcars)
   })
 }

 shinyApp(ui, server)
}
}

}
